pyproject.toml
[project]
name = "ash"
version = "0.1.0"
description = "Probabilistic price prediction model using next-token-prediction style architecture"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Deep learning
    "torch>=2.2.0",

    # Data processing
    "numpy>=1.26.0",
    "pandas>=2.2.0",

    # ONNX export
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
    "onnxscript>=0.1.0",

    # Configuration
    "pyyaml>=6.0",

    # Database connectivity
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "pandas-stubs>=2.2.0",
]

cuda = [
    # CUDA-enabled PyTorch (user installs separately via pip index)
    "onnxruntime-gpu>=1.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["data", "models", "evaluation"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
exclude = [
    "^inference-rs/",
]

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "onnx.*",
    "onnxruntime.*",
    "psycopg2.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["ash"]

[tool.pytest.ini_options]
testpaths = ["tests/python"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
